fix risk overflow and segfault

+++ CPP9/ex01/RPN.hpp
@@ -17,5 +17,5 @@ class RPN{
 
     private:
         RPN();
-        std::stack<long>     _value;
+        std::stack<float>     _value;
 };
